WE HERE.
5 AM.

+++ Baskel/src/BaskelBundle/Controller/EventController.php
@@ -5,6 +5,7 @@ namespace BaskelBundle\Controller;
 use BaskelBundle\Entity\Event;
 use BaskelBundle\Form\EventType;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
+use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 
 class EventController extends Controller
@@ -16,8 +17,18 @@ class EventController extends Controller
         $form->handleRequest($request);
         $em = $this->getDoctrine()->getManager();
         if ($form->isSubmitted()) {
+            /**
+             * @var UploadedFile $file
+             */
+            $file = $event->getImage();
+            $fileName = md5(uniqid()).'.'.$file->guessExtension();
+            $file->move(
+                $this->getParameter('image_directory'),$fileName
+            );
+            $event->setImage($fileName);
             $em->persist($event);
             $em->flush();
+            $this->addFlash("success","You have added an event successfully !");
             return $this->redirectToRoute('AfficheEv');
         }
         return $this->render('@Baskel/Default/NOCSSAjoutEvent.html.twig', array('f' => $form->createView()));
@@ -39,14 +50,20 @@ class EventController extends Controller
 }
 
 
-    function SupprimerEventAction($id)
+    function SupprimerEventAction($id,Request $request)
     {
-        $em = $this->getDoctrine()->getManager();
-        $event = $this->getDoctrine()->getRepository(Event::class)
-            ->find($id);
-        $em->remove($event);
-        $em->flush();
-        return $this->redirectToRoute('AfficheEv');
+        if($request->isXmlHttpRequest()) {
+            $id= $request->get('id');
+
+            $em = $this->getDoctrine()->getManager();
+            $event = $this->getDoctrine()->getRepository(Event::class)
+                ->find($id);
+            $em->remove($event);
+            $em->flush();
+
+            return new JsonResponse('good');
+
+        }
 
     }
